+++ src/utils/primes.h
@@ -4,6 +4,6 @@
 
 void genPrimes(int minRange, int maxRange, std::vector<int> *primeList);
 std::vector<int> multithreadPrimeGen(int maxNumber);
-uint64_t getRandomProduct(std::vector<uint64_t> *primes);
+uint64_t getRandomProduct(std::vector<uint32_t> *primes);
 
-std::vector<uint64_t> genPrimesGPU(void);
+std::vector<uint32_t> genPrimesGPU(void);
